[MINOR] move testGetRequiredSchema UT to TestHoodieSparkUtils
